+++ contracts/dutch-auction/CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Added
+
+## [0.1.2] - 2023-03-02
+### Added 
+- Add state consistency
 - New meta.
+### Changed
+- Updated `gstd`, `gtest`, `gear-wasm-builder` to `testnet`
 
 ## [0.1.1] - 2022-10-26
 ### Changed
@@ -16,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Initial release.
 
-[Unreleased]: https://github.com/gear-dapps/dutch-auction/compare/0.1.1...HEAD
+[Unreleased]: https://github.com/gear-dapps/dutch-auction/compare/0.1.2...HEAD
+[0.1.2]: https://github.com/gear-dapps/dutch-auction/compare/0.1.1...0.1.2
 [0.1.1]: https://github.com/gear-dapps/dutch-auction/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/gear-dapps/dutch-auction/compare/45b5e02...0.1.0